\name{medichiProfile}
\alias{medichiProfile.peaks}
\alias{medichiProfile.chip.deconv}
\alias{medichiProfile.deconv.entire.genome}
\title{Create MeDiChI profiles}

\description{
  Creates a profile generated by the MeDiChI model suitable for plotting. 
}

\usage{
  medichiProfile.peaks(peaks, kernel, by=1)
  medichiProfile.chip.deconv(fit, by=1)
  medichiProfile.deconv.entire.genome(fits, by=1)
}

\arguments{
  \item{peaks}{A 2-column matrix with position in the first column and intensity in the second.}
  \item{kernel}{A MeDiChI kernel.}
  \item{fit}{A MeDiChI chip.deconv object}
  \item{fits}{A MeDiChI chip.deconv.entire.genome object}
  \item{by}{Return every \emph{n}th row of the resulting profile}
}

\details{
  MeDiChI is a statistical model of hybridization of DNA fragments in a ChIP-chip experiment.
  A MeDiChI kernel is a 2-column matrix of positions and expected probe intensities around a
  site of protein-DNA binding, which can be tuned to an individual experiment.
  
  These functions work with MeDiChI output to produce plottable profiles. A profile is a fit
  of a curve to the MeDiChI model. By default the \strong{by} parameter is 1, which means the
  resulting matrix will give a predicted intensity at every base-pair. Since this can be a
  lot of data, particularly for whole-genome plots, using a larger \strong{by} parameter will
  yield a smaller profile matrix.
}

\value{
  A 2-column matrix with position in the first column and intensity in the second.
}

\references{
  \url{http://gaggle.systemsbiology.net/docs/geese/genomebrowser/},
  
  \url{http://baliga.systemsbiology.net/drupal/content/medichi}
}

\seealso{\code{\link{GenomeBrowserSupport-package}}}
\seealso{\code{\link{MeDiChI-package}}}
\seealso{\code{\link{chip.deconv}}}
\seealso{\code{\link{deconv.entire.genome}}}

\keyword{ track }
\keyword{ gaggle }
\keyword{ genome browser }
\keyword{ medichi }
\keyword{ profile }
